delete unnecessary code and console.logs
